[AUTO-68] chore: refactor schema files
