+++ src/server/api/router/dbRouter/students/mutation.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 import { createTRPCRouter, orgAdminOnlyPrecedure } from '@/server/api/trpc';
-import { studentFormSchema } from '@/base/schemas/student-form-schema';
+import { studentFormSchema } from '@/base/schemas/student-form.schema';
 
 export const mutationRouter = createTRPCRouter({
   add: orgAdminOnlyPrecedure
